# Removes allocations