```
benchmark                  old ns/op     new ns/op     delta
BenchmarkAgentSurfer-8     14217         12568         -11.60%

benchmark                  old allocs     new allocs     delta
BenchmarkAgentSurfer-8     11             4              -63.64%

benchmark                  old bytes     new bytes     delta
BenchmarkAgentSurfer-8     601           326           -45.76%
```